+++ x/spot/types/pool.go
@@ -83,6 +83,28 @@ func NewPool(
 }
 
 /*
+Ensure the denoms of the tokens in are assets of the pool
+
+args:
+
+  - tokensIn: the tokens to check
+
+ret:
+
+  - ok: true if all the denom from tokens in tokens in are in the pool
+*/
+func (pool Pool) AreTokensInDenomInPoolAssets(tokensIn sdk.Coins) bool {
+	for _, asset := range tokensIn {
+		_, _, err := pool.getPoolAssetAndIndex(asset.Denom)
+
+		if err != nil {
+			return false
+		}
+	}
+	return true
+}
+
+/*
 Adds tokens to a pool and updates the pool balances (i.e. liquidity).
 
 args:
@@ -192,7 +214,8 @@ func (pool *Pool) ExitPool(exitingShares sdk.Int) (
 	}
 
 	if !exitedCoins.IsValid() {
-		return sdk.Coins{}, errors.New("not enough pool shares to withdraw")
+		minShares := pool.MinSharesInForTokensOut()
+		return sdk.Coins{}, fmt.Errorf("not enough pool shares to withdraw - please provide at least %v shares", minShares)
 	}
 
 	// update the pool's balances